Move `DeleteDatasetIconButton` to its own file
