+++ web/src/features/datasets/components/DeleteDatasetButton.tsx
@@ -1,4 +1,3 @@
-import { IconOnlyButton } from "@/src/components/IconOnlyButton";
 import { Button, type ButtonProps } from "@/src/components/ui/button";
 import { DeleteDatasetDialog } from "@/src/features/datasets/components/DeleteDatasetDialog";
 import { usePostHogClientCapture } from "@/src/features/posthog-analytics/usePostHogClientCapture";
@@ -63,51 +62,3 @@ export const DeleteDatasetButton = forwardRef<
 });
 
 DeleteDatasetButton.displayName = "DeleteDatasetButton";
-
-export const DeleteDatasetIconButton = forwardRef<
-  HTMLButtonElement,
-  DeleteDatasetButtonProps
->((props, ref) => {
-  const capture = usePostHogClientCapture();
-  const [open, setOpen] = useState(false);
-  const hasAccess = useHasProjectAccess({
-    projectId: props.projectId,
-    scope: "datasets:CUD",
-  });
-
-  return (
-    <>
-      <IconOnlyButton
-        ref={ref}
-        icon={<Trash className="h-4 w-4" />}
-        label="Delete"
-        aria-label="delete"
-        disabledReason={
-          hasAccess
-            ? undefined
-            : "You don't have permission to delete this dataset."
-        }
-        variant={props.variant}
-        size={props.size}
-        className={props.className}
-        onClick={(event) => {
-          event.stopPropagation();
-          setOpen(true);
-          capture("datasets:delete_form_open", {
-            source: "table-single-row",
-          });
-        }}
-      />
-      <DeleteDatasetDialog
-        projectId={props.projectId}
-        datasetId={props.datasetId}
-        datasetName={props.datasetName}
-        open={hasAccess && open}
-        onOpenChange={setOpen}
-        trigger={{ type: "external" }}
-      />
-    </>
-  );
-});
-
-DeleteDatasetIconButton.displayName = "DeleteDatasetIconButton";
